% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{areaUnderIntensityRtQuantiles}
\alias{areaUnderIntensityRtQuantiles}
\title{Area Under Intensity-RT Quantiles}
\usage{
areaUnderIntensityRtQuantiles(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments}
}
\value{
`numeric(4)` named vector with Q1-Q4 area values
}
\description{
The function `areaUnderIntensityRtQuantiles` calculates the Area Under the
Curve (AUC) for 4 equal time intervals (retention time quartiles) across
the run.
}
\details{
The function splits the chromatogram into 4 equal time bins. To ensure accurate
area calculation at the boundaries, intensity values are linearly interpolated
exactly at the cut-off times.
}
\author{
Philippine Louail
}
