% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{areaUnderTicMs1}
\alias{areaUnderTicMs1}
\title{Area under TIC in MS1}
\usage{
areaUnderTicMs1(chromatograms, na.rm = TRUE, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{na.rm}{`logical(1)` whether to remove `NA` values (default `TRUE`)}

\item{...}{further arguments passed to `sum`}
}
\value{
`numeric(1)`
}
\description{
The function `areaUnderTicMs1` calculates the area under the total ion
current chromatogram for all MS1 spectra.
}
\details{
The function filters the chromatograms for `msLevel == 1` and calculates
the sum of intensities.

id: MS:4000029
name: area under TIC in MS1
def: "The area under the total ion current chromatogram (MS:1000235) of all MS1 spectra." [PSI:MS]
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 2L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0), intensity = c(100, 250, 400)),
    data.frame(rtime = c(5.1, 5.8), intensity = c(80, 500))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns sum of intensities for MS1 only
areaUnderTicMs1(chr)
}
\author{
Philippine Louail
}
