% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{areaUnderTicMs2}
\alias{areaUnderTicMs2}
\title{Area under TIC in MS2}
\usage{
areaUnderTicMs2(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `sum`}
}
\value{
`numeric(1)`
}
\description{
The function `areaUnderTicMs2` calculates the area under the total ion
current chromatogram for all MS2 spectra.
}
\details{
The function filters the chromatograms for `msLevel == 2` and calculates
the sum of intensities.

id: MS:4000030
name: area under TIC in MS2
def: "The area under the total ion current chromatogram (MS:1000235) of all MS2 spectra." [PSI:MS]
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 2L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0), intensity = c(100, 250, 400)),
    data.frame(rtime = c(5.1, 5.8), intensity = c(80, 500))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns sum of intensities for MS2 only
areaUnderTicMs2(chr)
}
\author{
Philippine Louail
}
