% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{chromatogramDuration}
\alias{chromatogramDuration}
\title{Calculate the duration of the chromatographic run per chromatogram}
\usage{
chromatogramDuration(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `min` and `max` (e.g. `na.rm`)}
}
\value{
`numeric(1)`
}
\description{
The function `chromatogramDuration` calculates The range of retention time
over which the data was acquired of the chromatographic run, defined as the
difference between the maximum and minimum retention time.
}
\details{
The function returns the duration of the chromatographic run.

id: MS:4000055
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L, 1L),
    mz = c(112.2, 123.3, 134.4),
    dataOrigin = c("mem1", "mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = numeric(), intensity = numeric()),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)

chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
chromatogramDuration(chr)
}
\author{
Philippine Louail
}
