% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{extentIntensity}
\alias{extentIntensity}
\title{Extent of Intensity across chromatograms}
\usage{
extentIntensity(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `min` and `max`}
}
\value{
`numeric(1)`
}
\description{
The function `extentIntensity` calculates the range of intensity values
(max - min) across all chromatograms.
}
\details{
The function returns `max - min` intensity across all chromatograms.

No specific PSI:MS term exists. It should be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns max - min intensity across all chromatograms: 1120
extentIntensity(chr)
}
\author{
Philippine Louail
}
