% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensity10xChange}
\alias{intensity10xChange}
\title{Number of 10x Intensity Changes across chromatograms}
\usage{
intensity10xChange(chromatograms, change = c("jump", "fall"), ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{change}{`character(1)` "jump" or "fall"}

\item{...}{further arguments}
}
\value{
`integer(1)`
}
\description{
The function `intensity10xChange` calculates the total number of times the
intensity changes by a factor of 10 or more between consecutive data points
across all chromatograms.
}
\details{
The function returns the total count of such changes across all chromatograms.

No specific PSI:MS term exists. It should be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(1.0, 2.0, 3.0, 4.0, 5.0),
               intensity = c(50, 100, 2000, 100, 50)),
    data.frame(rtime = c(1.0, 2.0, 3.0, 4.0, 5.0),
               intensity = c(200, 20, 30, 20, 200))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Count 10x intensity jumps (increases) across all chromatograms
intensity10xChange(chr, change = "jump")
## Count 10x intensity falls (decreases) across all chromatograms
intensity10xChange(chr, change = "fall")
}
\author{
Philippine Louail
}
