% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensityQuantileRtFraction}
\alias{intensityQuantileRtFraction}
\title{Fraction of Intensity in Equal RT Intervals}
\usage{
intensityQuantileRtFraction(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `range`}
}
\value{
`numeric(4)` named vector with fractions for Q1, Q2, Q3, Q4
}
\description{
The function `intensityQuantileRtFraction` calculates the fraction of the
total intensity that falls within 4 equal time intervals (retention time
quartiles) across the entire run.
}
\details{
The retention time range (min to max) is divided into 4 equal bins. The sum
of intensities for data points falling into each bin is calculated and
divided by the total intensity.
}
\author{
Philippine Louail
}
