% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensityRange}
\alias{intensityRange}
\title{Intensity Range per chromatogram}
\usage{
intensityRange(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `range`}
}
\value{
`numeric(2)` named vector with min and max
}
\description{
The function `intensityRange` calculates the range (min, max) of intensity
values for each chromatogram.
}
\details{
The function returns a list of numeric vectors of length 2.

No specific PSI:MS term exists. It should be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns (min, max) intensity across all chromatograms
intensityRange(chr)
}
\author{
Philippine Louail
}
