% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianPrecursorMz}
\alias{medianPrecursorMz}
\title{MS2 precursor median m/z of identified quantification data points
(MS:4000152)}
\usage{
medianPrecursorMz(
  spectra,
  msLevel = 1L,
  identificationLevel = c("all", "identified", "unidentified"),
  ...
)
}
\arguments{
\item{spectra}{\code{Spectra} object}

\item{msLevel}{\code{integer}}

\item{identificationLevel}{\code{character(1)}, one of \code{"all"},
\code{"identified"}, or \code{"unidentified"}}

\item{...}{not used here}
}
\value{
\code{numeric(1)}
}
\description{
MS:4000152 \cr
"Median m/z value for MS2 precursors of all quantification data points after
user-defined acceptance criteria are applied. These data points may be for
example XIC profiles, isotopic pattern areas, or reporter ions
(see MS:1001805). The used type should be noted in the metadata or
analysis methods section of the recording file for the respective run. In
case of multiple acceptance criteria (FDR) available in proteomics, PSM-level
FDR should be used for better comparability." [PSI:MS] \cr

The metric is calculated as follows: \cr
(1) the \code{Spectra} object is filtered according to the MS level, \cr
(2) the precursor m/z values are obtained, \cr
(3) the median value is returned (\code{NA}s are removed). \cr
}
\details{
MS:4000152 \cr
is_a: MS:4000003 ! single value \cr
is_a: MS:4000008 ! ID based \cr
is_a: MS:4000020 ! ion source metric \cr
relationship: has_metric_category MS:4000022 ! MS2 metric \cr
relationship: has_units MS:1000040 ! m/z

An attribute containing the PSI:MS term will only be returned if
\code{identificationLevel} is \code{"identified"} and \code{msLevel} is 1.
}
\note{
\code{medianPrecursorMz} will calculate the *precursor* median m/z of all
Spectra within \code{spectra}. If the calculation needs be done according to
*MS:4000152*, the \code{Spectra} object should be prepared accordingly, i.e.
filtered with e.g. [filterPrecursorMz()] or subsetted to spectra with
identification data.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
spd$precursorMz <- c(170.16, 170.16, 195.0876)
sps <- Spectra(spd)
medianPrecursorMz(spectra = sps, msLevel = 2L)
}
\author{
Thomas Naake
}
