\name{msQC-package}
\alias{msQC-package}
\alias{msQC}
\docType{package}
\title{
\packageTitle{msQC}
}
\description{
\packageDescription{msQC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{msQC}
\packageIndices{msQC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{msQC}

Maintainer: \packageMaintainer{msQC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
