% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_visualizations.R
\name{MsQuality}
\alias{MsQuality}
\title{Shiny application to visualize quality metrics}
\usage{
MsQuality(qc)
}
\arguments{
\item{qc}{`data.frame`}
}
\value{
`shiny`
}
\description{

}
\details{

}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
