% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{numberEmptyChrom}
\alias{numberEmptyChrom}
\title{Number of empty chromatograms}
\usage{
numberEmptyChrom(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments (currently ignored)}
}
\value{
`integer(1)`
}
\description{
The function `numberEmptyChrom` returns the number of chromatograms that are empty.
}
\details{
The function returns the number of empty chromatograms.

No specific PSI:MS term exists. It could be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L, 1L),
    mz = c(112.2, 123.3, 134.4),
    dataOrigin = c("mem1", "mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = numeric(), intensity = numeric()),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns 1 (the second chromatogram is empty)
numberEmptyChrom(chr)
}
\author{
Philippine Louail
}
