% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{rtAcquisitionRangeChromatograms}
\alias{rtAcquisitionRangeChromatograms}
\title{Retention time acquisition range}
\usage{
rtAcquisitionRangeChromatograms(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `range`}
}
\value{
`numeric(2)` named "min" and "max"
}
\description{
The function `rtAcquisitionRangeChromatograms` returns the minimum and
maximum retention time across all chromatograms.
}
\details{
The function returns a named vector of length 2 with the minimum and maximum
retention times.

id: MS:4000070
name: retention time acquisition range
def: "Upper and lower limit of retention time at which spectra are recorded." [PSI:MS]
is_a: MS:4000004 ! n-tuple
relationship: has_metric_category MS:4000009 ! ID free metric
relationship: has_metric_category MS:4000012 ! single run based metric
relationship: has_metric_category MS:4000016 ! retention time metric
relationship: has_units UO:0000010 ! second
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
rtAcquisitionRangeChromatograms(chr)
}
\author{
Philippine Louail
}
