% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtDuration}
\alias{rtDuration}
\title{RT duration (QC:4000053)}
\usage{
rtDuration(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"The retention time duration of the MS run in seconds, similar to the 
highest scan time minus the lowest scan time." [PSI:QC]
id: QC:4000053
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000021 ! retention time metric
}
\examples{
rtDuration(spectra)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
