% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtOverTICquantile}
\alias{rtOverTICquantile}
\title{RT over TIC quantile (QC:4000054)}
\usage{
rtOverTICquantile(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric(1)`}
}
\value{
`numeric(4)`
}
\description{
The interval when the respective quantile of the TIC accumulates divided by 
retention time duration. The number of quantiles observed is given by the 
size of the tuple." [PSI:QC]
id: QC:4000054
}
\details{
is_a: QC:4000004 ! n-tuple
is_a: QC:4000010 ! ID free
is_a: QC:4000021 ! retention time metric
is_a: QC:4000022 ! chromatogram metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
