% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{ticQuantileToQuantileLogRatio}
\alias{ticQuantileToQuantileLogRatio}
\title{MS1 quantile TIC change ratio to Quantile 1 (QC:4000057) or to 
previous quantile (QC:4000058)}
\usage{
ticQuantileToQuantileLogRatio(
  spectra,
  relativeTo = c("Q1", "previous"),
  MSLevel = 1L
)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{relativeTo}{`character`}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"The log ratio for the second to n-th quantile of TIC changes over first 
quantile of TIC changes." [PSI:QC]
id: QC:4000057
def: "The log ratio for the second to n-th quantile of TIC over the previous 
quantile of TIC. For the boundary elements min/max are used." [PSI:QC]
id: QC:4000058
}
\details{
is_a: QC:4000004 ! n-tuple
is_a: QC:4000010 ! ID free
is_a: QC:4000022 ! chromatogram metric
is_a: QC:4000023 ! MS1 metric

The `log2` values are returned instead of the `log` values.
}
\examples{
ticQuantileToQuantileLogRatio(spectra = spectra, relativeTo = "Q1",
    MSLevel = 1L)
ticQuantileToQuantileLogRatio(spectra, relativeTo = "previous",
    MSLevel = 1L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
