% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{xic50Fraction}
\alias{xic50Fraction}
\title{XIC50 fraction}
\usage{
xic50Fraction(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments}
}
\value{
`numeric(1)` fraction value
}
\description{
The function `xic50Fraction` calculates the fraction of XIC peaks that account
for the top half of all XIC-FWHM values.
}
\details{
This metric characterizes the distribution of precursor peak widths, similar
to how N50 characterizes contig sizes in genome assembly.

id: MS:4000050
name: XIC50 fraction
def: "The number of XIC that account for the top half of all XIC-FWHM divided
by the number of all XIC." [PSI:MS]
is_a: MS:4000003 ! single value
relationship: has_metric_category MS:4000009 ! ID free metric
relationship: has_metric_category MS:4000012 ! single run based metric
relationship: has_units UO:0000191 ! fraction
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L, 1L),
    mz = c(112.2, 123.3, 134.4),
    dataOrigin = c("mem1", "mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(4.0, 4.5, 5.0, 5.5, 6.0),
               intensity = c(50, 150, 300, 200, 75)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns fraction of XICs accounting for top half of FWHM
xic50Fraction(chr)
}
\author{
Philippine Louail
}
