% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/function_Chromatograms_metrics.R, R/function_Spectra_metrics.R
\name{areaUnderTic}
\alias{areaUnderTic}
\alias{areaUnderTic,Chromatograms-method}
\alias{areaUnderTIC,Spectra-method}
\alias{areaUnderTic,Spectra-method}
\title{Area under TIC}
\usage{
areaUnderTic(object, ...)

\S4method{areaUnderTic}{Chromatograms}(object, ...)

\S4method{areaUnderTic}{Spectra}(object, msLevel = 1L, ...)
}
\arguments{
\item{object}{\code{Spectra} or \code{Chromatograms} object}

\item{...}{additional arguments passed to internal helpers}

\item{msLevel}{\code{integer}}
}
\value{
`numeric(1)`

\code{numeric(1)}
}
\description{
The function `areaUnderTic` calculates the total area under the total ion
current chromatogram across all chromatograms.

MS:4000155 \cr
"The area under the total ion chromatogram." [PSI:MS] \cr

The metric is calculated as follows: \cr
(1) the input object is filtered according to the MS level, \cr
(2) the sum of the ion counts are obtained and returned.
}
\details{
Returns the sum of all intensity values across all chromatograms. Differences
between samples may indicate differences in dynamic range or sample content.

id: MS:4000155
name: area under TIC in MS1
def: "The area under the total ion current chromatogram (MS:1000235) of all
MS1 spectra." [PSI:MS]
comment: The metric informs about the dynamic range of the acquisition.
Differences between samples of an experiment may indicate differences in the
dynamic range and/or in the sample content. The signal can be affected by
contamination, retention time shifts, or loss of hydrophilic/hydrophobic
peptides.
is_a: MS:4000003 ! single value
relationship: has_metric_category MS:4000009 ! ID free metric
relationship: has_metric_category MS:4000017 ! chromatogram metric

MS:4000155 \cr
is_a: MS:4000003 ! single value \cr
relationship: has_metric_category MS:4000009 ! ID free metric \cr
relationship: has_metric_category MS:4000017 ! chromatogram metric \cr

The sum of the TIC is returned as an equivalent to the area. \cr
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns total area under TIC
areaUnderTic(chr)
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
sps <- Spectra(spd)
areaUnderTic(object = sps, msLevel = 2L)
}
\author{
Philippine Louail

Thomas Naake
}
