% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{areaUnderTic}
\alias{areaUnderTic}
\alias{areaUnderTic,Chromatograms-method}
\alias{areaUnderTic,Spectra-method}
\title{area under TIC (MS:4000155)}
\usage{
areaUnderTic(object, msLevel = 1L, na.rm = TRUE, ...)
}
\arguments{
\item{object}{\code{Spectra} or \code{Chromatograms} object}

\item{msLevel}{\code{integer}}

\item{na.rm}{\code{logical(1)} whether to remove \code{NA} values before
computing the sum (default \code{TRUE})}

\item{...}{additional arguments passed to internal helpers}
}
\value{
For \code{Spectra}: \code{numeric(1)}.
    For \code{Chromatograms}: \code{numeric} of length equal to
    \code{length(object)}, one area per chromatogram.
}
\description{
MS:4000155 \cr
"The area under the total ion chromatogram." [PSI:MS] \cr

The metric is calculated as follows: \cr
(1) the input object is filtered according to the MS level, \cr
(2) the sum of the ion counts are obtained and returned
(\code{NA} values are removed by default, controlled by \code{na.rm}).
}
\details{
MS:4000155 \cr
is_a: MS:4000003 ! single value \cr
relationship: has_metric_category MS:4000009 ! ID free metric \cr
relationship: has_metric_category MS:4000017 ! chromatogram metric \cr

The sum of the TIC is returned as an equivalent to the area. \cr
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
sps <- Spectra(spd)
areaUnderTic(object = sps, msLevel = 2L)
}
\author{
Thomas Naake
}
