% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{areaUnderTIC}
\alias{areaUnderTIC}
\title{Area under TIC (QC:4000077)}
\usage{
areaUnderTIC(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"The area under the total ion chromatogram." [PSI:QC]
id: QC:4000077
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000022 ! chromatogram metric

The sum of the TIC is returned as an equivalent to the area.
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
