% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{areaUnderTICRTquantiles}
\alias{areaUnderTICRTquantiles}
\title{Area under TIC RT quantiles (QC:4000078)}
\usage{
areaUnderTICRTquantiles(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(4)`
}
\description{
"The area under the total ion chromatogram of the retention time quantiles. 
Number of quantiles are given by the n-tuple." [PSI:QC]
id: QC:4000078
}
\details{
is_a: QC:4000004 ! n-tuple
is_a: QC:4000010 ! ID free
is_a: QC:4000022 ! chromatogram metric

The sum of the TIC is returned as an equivalent to the area.
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
