% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/function_calculateMetrics_from_Chromatograms.R,
%   R/function_calculateMetrics_from_Spectra.R
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,Chromatograms-method}
\alias{calculateMetrics,Spectra-method}
\alias{calculateMetrics,MsExperiment-method}
\title{Calculate QC metrics from a Spectra, MsExperiment, or Chromatograms object}
\usage{
calculateMetrics(object, metrics, ...)

\S4method{calculateMetrics}{Chromatograms}(
  object,
  metrics = qualityMetrics(object),
  filterEmptyObject = FALSE,
  ...
)

\S4method{calculateMetrics}{Spectra}(
  object,
  metrics = qualityMetrics(object),
  filterEmptyObject = FALSE,
  ...
)

\S4method{calculateMetrics}{MsExperiment}(
  object,
  metrics = qualityMetrics(object),
  filterEmptyObject = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{Spectra}, \code{MsExperiment}, or \code{Chromatograms} object}

\item{metrics}{\code{character} specifying the quality metrics to be
calculated on \code{object}}

\item{...}{arguments passed to the quality metrics functions defined in
\code{metrics}}

\item{filterEmptyObject}{\code{logical(1)} specifying if empty entries and
entries with intensity zero of the \code{Spectra} object will be removed}
}
\value{
\code{data.frame} containing in the columns the metrics for the
different spectra/chromatograms and in rows the samples
}
\description{
Calculate QC metrics from a \code{Spectra}, \code{MsExperiment}, or
\code{Chromatograms} object.
\code{calculateMetrics} is a generic function that dispatches to the
appropriate method based on the class of the input object.
}
\details{
The metrics are defined by the argument \code{metrics}. Further arguments
passed to the quality metric functions can be specified by the \code{params}
argument. \code{params} can contain named entries which are matched against
the formal arguments of the quality metric functions.

Setting the argument \code{filterEmptyObject} to \code{TRUE} will remove
zero-length entries, zero-intensity entries, and entries with
intensities that are \code{Inf} from the \code{Spectra} object.

For \code{Chromatograms} objects, the argument \code{filterEmptyObject}
is interpreted as \code{filterEmptyChromatograms}.
}
\examples{
library(MsDataHub)
library(Spectra)

## define file names containing spectra data for the samples
sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
   MsDataHub::X20171016_POOL_POS_3_105.134.mzML())

## import the data and assign it to the spectra object
spectra <- Spectra(sciex_file)

## define the quality metrics to be calculated
metrics <- c("areaUnderTic", "chromatographyDuration", "msSignal10xChange")

## calculate the metrics
## additional parameters passed to the quality metrics functions
## (MsLevel is an argument of areaUnderTic and msSignal10xChange,
## relativeTo is an argument of msSignal10xChange) passed to ...
calculateMetrics(object = spectra, metrics = metrics,
    msLevel = 1, change = "jump", relativeTo = "Q1")
calculateMetrics(object = spectra, metrics = metrics,
    msLevel = 1, change = "fall", relativeTo = "previous")
}
\author{
Thomas Naake
}
