% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calculateMetrics_from_Chromatograms.R
\name{calculateMetricsFromChromatograms}
\alias{calculateMetricsFromChromatograms}
\title{Calculate QC metrics from a Chromatograms object}
\usage{
calculateMetricsFromChromatograms(
  chromatograms,
  metrics,
  filterEmptyObject = FALSE,
  f = dataOrigin(chromatograms),
  format = c("data.frame", "mzQC"),
  ...,
  BPPARAM = bpparam()
)
}
\arguments{
\item{chromatograms}{\code{Chromatograms} object}

\item{metrics}{\code{character} specifying the quality metrics to be
calculated on \code{chromatograms}}

\item{filterEmptyObject}{\code{logical(1)} specifying if empty entries
and entries with intensity zero of the \code{Chromatograms} object will be
removed}

\item{f}{\code{character} defining which chromatograms in \code{chromatograms}
belong to one sample. Defaults to \code{f = dataOrigin(chromatograms)}.
Chromatograms from the same original data file are processed together
(and in parallel for different files).}

\item{format}{\code{character(1)} specifying if metrics are returned
as a \code{data.frame} (\code{format = "data.frame"}) or as a list of
\code{MzQCmzQC} objects (\code{format = "mzQC"})}

\item{...}{arguments passed to the quality metrics functions defined in
\code{metrics}}

\item{BPPARAM}{Parallel processing setup. Defaults to \code{BPPARAM = bpparam()}.
See [bpparam()] for details on parallel processing with \code{BiocParallel}.}
}
\value{
In case of \code{format = "data.frame"}, a \code{data.frame} containing in
the columns the metrics for the different chromatograms of identical
\code{dataOrigin{chromatograms}} (in rows).
In case of \code{format = "mzQC"}, a \code{list} of \code{MzQCmzQC} objects
containing the metrics for the different chromatograms of identical
\code{dataOrigin{chromatograms}}
}
\description{
The function \code{calculateMetricsFromChromatograms} calculates quality
metrics from a \code{Chromatograms} object. The function will calculate the
metrics per sample according to the grouping parameter \code{f},
e.g. \code{dataOrigin} information.
}
\details{
The metrics are defined by the argument \code{metrics}. Further arguments
passed to the quality metric functions can be specified by \code{...}.
The additional arguments \code{...} are matched against
the formal arguments of the quality metric functions.
Samples will be processed in parallel using the default parallel processing
setup ([bpparam()]) or with the parallel processing setup defined with
parameter \code{BPPARAM}.

Setting the argument \code{filterEmptyObject} to \code{TRUE} will
remove zero-length entries, zero-intensity entries, and entries with
intensities that are \code{Inf} from the \code{Chromatograms} object.
}
\examples{
library(Chromatograms)
## Create a Chromatograms object with ChromBackendMemory
cdata <- data.frame(
    msLevel = c(1L, 1L, 1L),
    mz = c(112.2, 123.3, 134.4),
    dataOrigin = c("mem1", "mem2", "mem3")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(3.5, 4.0, 4.5),
               intensity = c(80, 120, 90)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
metrics <- c("chromatogramDuration", "maxIntensity", "intensityMean")
}
\author{
Philippine Louail
}
