% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calculateMetrics_from_Spectra.R
\name{calculateMetricsFromSpectra}
\alias{calculateMetricsFromSpectra}
\title{Calculate QC metrics from a Spectra object}
\usage{
calculateMetricsFromSpectra(
  spectra,
  metrics = qualityMetrics(spectra),
  f = dataOrigin(spectra),
  ...,
  BPPARAM = bpparam()
)
}
\arguments{
\item{spectra}{\code{Spectra} object}

\item{metrics}{\code{character} specifying the quality metrics to be 
calculated on \code{spectra}}

\item{f}{\code{character} defining which spectra in `spectra` belong to
one sample. Defaults to `f = dataOrigin(spectra)`. Spectra from the
same original data file are processed together (and in parallel for
different files).}

\item{...}{arguments passed to the quality metrics functions defined in 
\code{metrics}}

\item{BPPARAM}{Parallel processing setup. Defaults to `BPPARAM = bpparam()`.
See [bpparam()] for details on parallel processing with `BiocParallel`.}
}
\value{
\code{data.frame} containing in the columns the metrics for the 
different spectra (in rows)
}
\description{
The function \code{calculateMetricsFromSpectra} calculates quality metrics 
from a \code{Spectra} object. The function will calculate the 
metrics per sample according to the grouping parameter \code{f}, 
e.g. \code{dataOrigin} information. Samples will be processed in parallel
using the default parallel processing setup ([bpparam()]) or with the
parallel processing setup defined with parameter `BPPARAM`.
}
\details{
The metrics are defined by the argument \code{metrics}. Further arguments 
passed to the quality metric functions can be specified by the \code{params}
argument. \code{params} can contain named entries which are matched against 
the formal arguments of the quality metric functions.
}
\examples{
library(msdata)
library(Spectra)

## define file names containing spectra data for the samples
fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)

## import the data and add it to the spectra object
spectra <- Spectra(fls, backend = MsBackendMzR())

## define the quality metrics to be calculated
metrics <- c("areaUnderTic", "rtDuration", "msSignal10xChange")

## calculate the metrics
## additional parameters passed to the quality metrics functions
## (msLevel is an argument of areaUnderTic and msSignal10xChange,
## relativeTo is an argument of msSignal10xChange) passed to ...
calculateMetricsFromSpectra(spectra = spectra, metrics = metrics, 
    msLevel = 1, change = "jump", relativeTo = "Q1")
calculateMetricsFromSpectra(spectra = spectra, metrics = metrics, 
    msLevel = 1, change = "fall", relativeTo = "previous")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}, Johannes Rainer
}
