% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{chromatogramCount}
\alias{chromatogramCount}
\title{Number of Chromatograms}
\usage{
chromatogramCount(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments}
}
\value{
`integer(1)`
}
\description{
The function `chromatogramCount` computes the number of chromatograms in the
`Chromatograms` object.
}
\details{
This metric corresponds to the PSI:MS term:
MS:4000071 (number of chromatograms).
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns number of chromatograms
chromatogramCount(chr)
}
\author{
Philippine Louail
}
