% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{extentIdentifiedPrecursorIntensity}
\alias{extentIdentifiedPrecursorIntensity}
\title{Extent of identified precursor intensity (QC:4000125)}
\usage{
extentIdentifiedPrecursorIntensity(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"Ratio of 95th over 5th percentile of precursor intensity for identified 
peptides" [PSI:QC]
id: QC:4000125
}
\details{
Can be used to approximate the dynamic range of signal
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
