% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensityQuartiles}
\alias{intensityQuartiles}
\title{Intensity Quartiles per chromatogram}
\usage{
intensityQuartiles(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments passed to `summary`}
}
\value{
`numeric(6)` named vector with Min, 1st Qu., Median, Mean, 3rd Qu., Max
}
\description{
The function `intensityQuartiles` calculates the minimum, 1st quartile,
median, 3rd quartile, and maximum of intensity values for each chromatogram.
}
\details{
The function returns a list of vectors, where each vector contains the
summary statistics for a chromatogram.

No specific PSI:MS term exists. It should be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns Min, 1st Qu., Median, Mean, 3rd Qu., Max for the overall intensities
intensityQuartiles(chr)
}
\author{
Philippine Louail
}
