% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensitySd}
\alias{intensitySd}
\title{Intensity statistics (Standard Deviation) across chromatograms}
\usage{
intensitySd(chromatograms, na.rm = TRUE, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{na.rm}{`logical(1)` whether to remove `NA` values (default `TRUE`)}

\item{...}{further arguments passed to `sd`}
}
\value{
`numeric(1)`
}
\description{
The function `intensitySd` calculates the standard deviation of intensity
values across all chromatograms.
}
\details{
The function returns the standard deviation of the intensity across all chromatograms.

No specific PSI:MS term exists. It should be created.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns standard deviation of intensities across all chromatograms
intensitySd(chr)
}
\author{
Philippine Louail
}
