% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{intensitySd}
\alias{intensitySd}
\title{Intensity statistics (Standard Deviation) per chromatogram}
\usage{
intensitySd(chromatograms, msLevel = integer(), na.rm = TRUE, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{msLevel}{`integer` defining the MS level(s) to filter for.
If empty (default), no filtering is performed.}

\item{na.rm}{`logical(1)` whether to remove `NA` values (default `TRUE`)}

\item{...}{further arguments passed to `sd`}
}
\value{
`numeric` of length equal to `length(chromatograms)`
}
\description{
The function `intensitySd` calculates the standard deviation of intensity
values within each chromatogram. \cr

The metric is calculated as follows: \cr
(1) for each chromatogram in the `Chromatograms` object, the intensity
values are extracted, \cr
(2) the standard deviation per chromatogram is calculated and returned.
}
\examples{
library(Chromatograms)
cdata <- data.frame(
    msLevel = c(1L, 1L),
    mz = c(112.2, 123.3),
    dataOrigin = c("mem1", "mem1")
)
pdata <- list(
    data.frame(rtime = c(2.1, 2.5, 3.0, 3.4, 3.9),
               intensity = c(100, 250, 400, 300, 150)),
    data.frame(rtime = c(5.1, 5.8, 6.3, 6.9, 7.5),
               intensity = c(80, 500, 1200, 600, 120))
)
chr <- Chromatograms(ChromBackendMemory(), chromData = cdata, peaksData = pdata)
## Returns standard deviation of intensities per chromatogram
intensitySd(chr)
}
\author{
Philippine Louail
}
