% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{meanCharge}
\alias{meanCharge}
\title{Mean charge in identified spectra (QC:4000177) or mean precursor 
charge in all MS2 (QC:4000182)}
\usage{
meanCharge(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Mean charge in identified spectra" [PSI:QC]
id: QC:4000177

"Mean precursor charge in all MS2" [PSI:QC]
id: QC:4000182
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
