% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianCharge}
\alias{medianCharge}
\title{Median charge in identified spectra (QC:4000178) or median
precursor charge in all MS2 (QC:4000183)}
\usage{
medianCharge(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Median charge in identified spectra" [PSI:QC]
id: QC:4000178

"Median precursor charge in all MS2" [PSI:QC]
id: QC:4000183
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
