% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianCharge}
\alias{medianCharge}
\title{Median charge in identified spectra (QC:4000178) or median
precursor charge in all MS2 (QC:4000183)}
\usage{
medianCharge(spectra, msLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{msLevel}{`integer(1)`}
}
\value{
`numeric(1)`
}
\description{
"Median charge in identified spectra" [PSI:QC]
id: QC:4000178

"Median precursor charge in all MS2" [PSI:QC]
id: QC:4000183

The metric is calculated as follows:
(1) the `Spectra` object is filtered according to the MS level,

(2) the precursor charge is obtained,

(3) the median of the precursor charge values is calculated and returned.
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\note{
The `Spectra` object might contain features that were not identified. If
the calculation needs to be done according to *QC:4000178*, the 
`Spectra` object should be prepared accordingly.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
sps <- Spectra(spd)
spd$precursorCharge <- c(1L, 1L, 1L)
medianCharge(spectra = sps, msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
