% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianPrecursorMZ}
\alias{medianPrecursorMZ}
\title{Precursor median m/z for IDs (QC:4000065)}
\usage{
medianPrecursorMZ(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Median m/z value for all identified peptides (unique ions) after 
FDR." [PSI:QC]
id: QC:4000065
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000023 ! MS1 metric
is_a: QC:4000025 ! ion source metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
