% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianPrecursorMZ}
\alias{medianPrecursorMZ}
\title{Precursor median m/z for IDs (QC:4000065)}
\usage{
medianPrecursorMZ(spectra, msLevel = 1L, ...)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{msLevel}{`integer`}

\item{...}{not used here}
}
\value{
`numeric(1)`
}
\description{
"Median m/z value for all identified peptides (unique ions) after 
FDR." [PSI:QC]
id: QC:4000065

The metric is calculated as follows:
(1) the `Spectra` object is filtered according to the MS level,

(2) the precursor m/z values are obtained,

(3) the median value is returned (`NAs` are removed).
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000023 ! MS1 metric
is_a: QC:4000025 ! ion source metric
}
\note{
`medianPrecursorMZ` will calculate the *precursor* median m/z of all 
Spectra within `spectra`. If the calculation needs be done according to
*QC:4000065*, the `Spectra` object should be prepared accordingly, i.e.
filtered with e.g. [filterPrecursorMz()] or subsetted to spectra with
identification data.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
spd$precursorMz <- c(170.16, 170.16, 195.0876)
sps <- Spectra(spd)
medianPrecursorMZ(spectra = sps, msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
