% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianTICofRTRange}
\alias{medianTICofRTRange}
\title{Median of TIC values in the shortest RT range in which half of the 
peptides are identified (QC:4000132)}
\usage{
medianTICofRTRange(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"Median of TIC values in the shortest RT range in which half of the 
peptides are identified"  [PSI:QC] 
id: QC:4000132
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
