% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{medianTICRTIQR}
\alias{medianTICRTIQR}
\title{Median of TIC values in the RT range in which the middle half of 
peptides are identified (QC:4000130)}
\usage{
medianTICRTIQR(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Median of TIC values in the RT range in which half of peptides are 
identified (RT values of Q1 to Q3 of identifications)" [PSI:QC]
id: QC:4000130
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric

The function `medianTICrtIQR` uses the function `ionCount` as an 
equivalent to the TIC.
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
