\name{msQC-package}
\alias{msQC-package}
\alias{msQC}
\docType{package}
\title{
\packageTitle{msQC}
}
\description{
msQC enables to calculate quality metrics of mass spectrometry
data. It is build upon Spectra and MsExperiment objects.
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{msQC}
\packageIndices{msQC}
}

\author{
\packageAuthor{msQC}
Maintainer: \packageMaintainer{msQC}
}

\keyword{ package, QC, proteomics, metabolomics, mass spectrometry }

\examples{
\dontrun{calculateMetrics(object = spectra)}
\dontrun{calculateMetrics(object = mse)}
}
