\name{MsQuality-package}
\alias{MsQuality-package}
\alias{MsQuality}
\docType{package}
\title{
\packageTitle{MsQuality}
}
\description{
MsQuality enables to calculate quality metrics of mass spectrometry
data. It is build upon Spectra and MsExperiment objects.
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{MsQuality}
\packageIndices{MsQuality}
}

\author{
\packageAuthor{MsQuality}
Maintainer: \packageMaintainer{MsQuality}
}

\keyword{ package, QC, proteomics, metabolomics, mass spectrometry }

\examples{
\dontrun{calculateMetrics(object = spectra)}
\dontrun{calculateMetrics(object = mse)}
}
