% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{msSignal10XChange}
\alias{msSignal10XChange}
\title{MS1 signal jump/fall (10x) count (QC:4000172/QC:400173)}
\usage{
msSignal10XChange(spectra, change = c("jump", "fall"), MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{change}{`character`, one of `"jump"` or `"fall"`}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"The count of MS1 signal jump (spectra sum) by a factor of ten or more (10x)
between two subsequent scans" [PSI:QC]
id: QC:4000172

"The count of MS1 signal decline (spectra sum) by a factor of ten or more 
(10x) between two subsequent scans" [PSI:QC]
id: QC:4000173
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric
An unusual high count of signal jumps or falls can indicate ESI stability 
issues.
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
