% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{msSignal10XChange}
\alias{msSignal10XChange}
\title{MS1 signal jump/fall (10x) count (QC:4000172/QC:4000173)}
\usage{
msSignal10XChange(spectra, change = c("jump", "fall"), msLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{change}{`character`, one of `"jump"` or `"fall"`}

\item{msLevel}{`integer(1)`}
}
\value{
`numeric(1)`
}
\description{
"The count of MS1 signal jump (spectra sum) by a factor of ten or more (10x)
between two subsequent scans" [PSI:QC]
id: QC:4000172

"The count of MS1 signal decline (spectra sum) by a factor of ten or more 
(10x) between two subsequent scans" [PSI:QC]
id: QC:4000173

#' The metric is calculated as follows:
(1) the `Spectra` object is filtered according to the MS level,

(2) the intensity of the precursor ions within `spectra` are obtained,

(3) the intensity values of the features are obtained via the ion count,

(4) the signal jumps/declines of the intensity values with the two 
subsequent intensity values is calculated,

(5) in the case of *QC:4000172*, the signal jumps by a factor of ten or more
are counted and returned;
in the case of *QC:4000173*, the signal declines by a factor of ten or more
are counted and returned.
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric

An unusual high count of signal jumps or falls can indicate ESI stability 
issues.

The function `msSignal10XChange` uses the function `ionCount` as an 
equivalent to the TIC.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
sps <- Spectra(spd)
msSignal10XChange(spectra = sps, change = "jump", msLevel = 2L)
msSignal10XChange(spectra = sps, change = "fall", msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
