% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{mzAcquisitionRange}
\alias{mzAcquisitionRange}
\title{m/z acquisition range (QC:4000138)}
\usage{
mzAcquisitionRange(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(2)`
}
\description{
"Upper and lower limit of m/z values at which spectra are recorded." [PSI:QC]
id: QC:4000138
}
\details{
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric
is_a: QC:4000004 ! n-tuple
}
\examples{
mzAcquisitionRange(spectra)

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
