% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{numberSpectra}
\alias{numberSpectra}
\title{Number of MS1/MS2 spectra (QC:4000059/QC:4000060)}
\usage{
numberSpectra(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric(1)`}
}
\value{
`numeric(1)`
}
\description{
"The number of MS1 events in the run." [PSI:QC]
id: QC:4000059
"The number of MS2 events in the run." [PSI:QC]
id: QC:4000060
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000023 ! MS1 metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
