% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{precursorIntensityMean}
\alias{precursorIntensityMean}
\title{Precursor intensity distribution mean (QC:4000168),
Identified precursor intensity distribution mean (QC:4000229), or
Unidentified precursor intensity distribution mean (QC:4000234)}
\usage{
precursorIntensityMean(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"From the distribution of precursor intensities, the mean." [PSI:QC]
id: QC:4000168

"From the distribution of identified precursor intensities, the mean"
 [PSI:QC]
id: QC:4000229

"From the distribution of unidentified precursor intensities, the mean" 
[PSI:QC]
id: QC:4000234
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric

The intensity distribution of the precursors informs about the dynamic 
range of the acquisition.

The intensity distribution of the identified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.

The intensity distribution of the unidentified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.
}
\examples{
precursorIntensityMean(spectra)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
