% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{precursorIntensityQuartiles}
\alias{precursorIntensityQuartiles}
\title{Precursor intensity distribution Q1, Q2, Q3 (QC:4000167), 
Identified precursor intensity distribution Q1, Q2, Q3 (QC:4000228), or
Unidentified precursor intensity distribution Q1, Q2, Q3 (QC:4000233)}
\usage{
precursorIntensityQuartiles(spectra, MSLevel = 1L)
}
\arguments{
\item{object}{`MSnExp` or `Spectra` object}
}
\value{
`numeric(3)`
}
\description{
"From the distribution of precursor intensities, the quartiles Q1, Q2, Q3" 
[PSI:QC]
id: QC:4000167

"From the distribution of identified precursor intensities, the quartiles 
Q1, Q2, Q3" [PSI:QC]
id: QC:40000228

"From the distribution of unidentified precursor intensities, the 
quartiles Q1, Q2, Q3" [PSI:QC]
id: QC:4000233
}
\details{
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric
is_a: QC:4000004 ! n-tuple

The intensity distribution of the precursors informs about the dynamic 
range of the acquisition.

The intensity distribution of the identified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.

The intensity distribution of the unidentified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.
}
\examples{
precursorIntensityQuartiles(spectra)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
