% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{precursorIntensityRange}
\alias{precursorIntensityRange}
\title{Precursor intensity range (QC:4000144)}
\usage{
precursorIntensityRange(spectra, MSLevel = 1)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(2)`
}
\description{
"Minimum and maximum precursor intensity recorded." [PSI:QC]
id: QC:4000144
}
\details{
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric
is_a: QC:4000004 ! n-tuple

The intensity range of the precursors informs about the dynamic range of 
the acquisition.
}
\examples{


}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
