% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{precursorIntensitySD}
\alias{precursorIntensitySD}
\title{Precursor intensity distribution sigma (QC:4000169),
Identified precursor intensity distribution sigma (QC:4000230), or 
Unidentified precursor intensity distribution sigma (QC:4000235)}
\usage{
precursorIntensitySD(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"From the distribution of precursor intensities, the sigma value." [PSI:QC]
id: QC:4000169

"From the distribution of identified precursor intensities, the sigma 
value" [PSI:QC]
id: QC:4000230

"From the distribution of unidentified precursor intensities, the 
sigma value" [PSI:QC]
id: QC:4000235
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric

The intensity distribution of the precursors informs about the dynamic 
range of the acquisition.

The intensity distribution of the identified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.

The intensity distribution of the unidentified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.
}
\examples{
precursorIntensitySD(spectra)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
