% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{precursorIntensitySd}
\alias{precursorIntensitySd}
\title{Precursor intensity distribution sigma (QC:4000169),
Identified precursor intensity distribution sigma (QC:4000230), or 
Unidentified precursor intensity distribution sigma (QC:4000235)}
\usage{
precursorIntensitySd(spectra, msLevel = 1L, ...)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{msLevel}{`integer`}

\item{...}{not used here}
}
\value{
`numeric(1)`
}
\description{
"From the distribution of precursor intensities, the sigma value." [PSI:QC]
id: QC:4000169

"From the distribution of identified precursor intensities, the sigma 
value" [PSI:QC]
id: QC:4000230

"From the distribution of unidentified precursor intensities, the 
sigma value" [PSI:QC]
id: QC:4000235

The metric is calculated as follows:
(1) the `Spectra` object is filtered according to the MS level,

(2) the intensity of the precursor ions within `spectra` are obtained,

(3) the standard deviation of precursor intensity values is obtained 
(`NA` values are removed) and returned.
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000010 ! ID free
is_a: QC:4000001 ! QC metric

The intensity distribution of the precursors informs about the dynamic 
range of the acquisition.

The intensity distribution of the identified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.

The intensity distribution of the unidentified precursors informs about the 
dynamic range of the acquisition in relation to identifiability.
}
\note{
The `Spectra` object might contain features that were (not) identified. If
the calculation needs to be done according to *QC:4000230*/*QC:4000235*, the 
`Spectra` object should be prepared accordingly.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
spd$precursorIntensity <- c(100.0, 100.0, 100.0)
sps <- Spectra(spd)
precursorIntensitySd(spectra = sps, msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
