% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_or_MsExperiment_metrics.R
\name{qualityMetrics}
\alias{qualityMetrics}
\title{Get a vector of quality metrics than can be applied to `object`}
\usage{
qualityMetrics(object)
}
\arguments{
\item{object}{object of type `Spectra` or `MsExperiment`}
}
\value{
`character`
}
\description{
The function `qualityMetrics` returns a character vector with available
quality metrics depending on `object`.
}
\details{
`object` is either a `Spectra` or `MsExperiment` object.
}
\examples{
library(msdata)
library(Spectra)
fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
spectra <- Spectra(fls, backend = MsBackendMzR())

qualityMetrics(spectra)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
