% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{RatioCharge1over2}
\alias{RatioCharge1over2}
\title{Charged peptides ratio 1+ over 2+ (QC:4000174) or 
Charged spectra ratio +1 over +2 (QC:4000179)}
\usage{
RatioCharge1over2(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Ratio of 1+ peptide count over 2+ peptide count in identified spectra" [PSI:QC]
id: QC:4000174

"Ratio of 1+ spectra count over 2+ spectra count in all MS2" [PSI:QC]
id: QC:4000179
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
