% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{RatioCharge1over2}
\alias{RatioCharge1over2}
\title{Charged peptides ratio 1+ over 2+ (QC:4000174) or 
Charged spectra ratio 1+ over 2+ (QC:4000179)}
\usage{
RatioCharge1over2(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"Ratio of 1+ peptide count over 2+ peptide count in identified spectra" [PSI:QC]
id: QC:4000174

"Ratio of 1+ spectra count over 2+ spectra count in all MS2" [PSI:QC]
id: QC:4000179
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
spd$precursorCharge <- c(1L, 1L, 1L)
sps <- Spectra(spd)
RatioCharge1over2(spectra = sps, MSLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
