% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{RatioCharge3over2}
\alias{RatioCharge3over2}
\title{Charged peptides ratio 3+ over 2+ (QC:4000175) or 
charged spectra ratio +3 over +2 (QC:4000180)}
\usage{
RatioCharge3over2(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Ratio of 3+ peptide count over 2+ peptide count in identified spectra" [PSI:QC]
id: QC:4000175

"Ratio of 3+ peptide count over 2+ peptide count in all MS2" [PSI:QC]
id: QC:4000180
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
