% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{RatioCharge4over2}
\alias{RatioCharge4over2}
\title{Charged peptides ratio 4+ over 2+ (QC:4000176) or charged spectra 
ratio 4+ over 2+ (QC:4000181)}
\usage{
RatioCharge4over2(spectra)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(1)`
}
\description{
"Ratio of 4+ peptide count  over 2+ peptide count in identified 
spectra" [PSI:QC]
id: QC:4000176

"Ratio of 4+ peptide count over 2+ peptide count in all MS2" [PSI:QC]
id: QC:4000181
}
\details{
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000001 ! QC metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
