% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtAcquisitionRange}
\alias{rtAcquisitionRange}
\title{Retention time acquisition range (QC:4000139)}
\usage{
rtAcquisitionRange(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}
}
\value{
`numeric(2)`
}
\description{
"Upper and lower limit of time at which spectra are recorded." [PSI:QC]
id: QC:4000139
}
\details{
is_a: QC:4000004 ! n-tuple
}
\examples{
rtAcquisitionRange(spectra)

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
