% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtIQR}
\alias{rtIQR}
\title{Interquartile RT period for peptide identifications (QC:4000072)}
\usage{
rtIQR(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(1)`
}
\description{
"The interquartile retention time period, in seconds, for all peptide 
identifications over the complete run." [PSI:QC]
id: QC:4000072
}
\details{
Longer times indicate better chromatographic separation.
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000022 ! chromatogram metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
