% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtIqr}
\alias{rtIqr}
\title{Interquartile RT period for peptide identifications (QC:4000072)}
\usage{
rtIqr(spectra, msLevel = 1L, ...)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{msLevel}{`integer`}

\item{...}{not used here}
}
\value{
`numeric(1)`
}
\description{
"The interquartile retention time period, in seconds, for all peptide 
identifications over the complete run." [PSI:QC]
id: QC:4000072

The metric is calculated as follows:
(1) the `Spectra` object is filtered according to the MS level,

(2) the retention time values are obtained,

(3) the interquartile range is obtained from the values and returned
(`NA` values are removed).
}
\details{
Longer times indicate better chromatographic separation.
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000022 ! chromatogram metric

Retention time values that are `NA` are removed.
}
\note{
The `Spectra` object might contain features that were not identified. If
the calculation needs to be done according to *QC:4000072*, the 
`Spectra` object should be prepared accordingly, i.e. subsetted to spectra
with identification data.

The stored retention time information in `spectra` might have a different
unit than seconds. `rtIqr` will return the IQR based on the values stored
in `spectra` and will not convert these values to seconds.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
spd$rtime <- c(9.44, 9.44, 15.84)
sps <- Spectra(spd)
rtIqr(spectra = sps, msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
