% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtIQRrate}
\alias{rtIQRrate}
\title{Peptide identification rate of the interquartile RT period (QC:4000073)}
\usage{
rtIQRrate(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric`}
}
\value{
`numeric(2)`
}
\description{
The identification rate of peptides for the interquartile retention time 
period, in peptides per second." [PSI:QC]
id: QC:4000073
}
\details{
Higher rates indicate efficient sampling and identification.
is_a: QC:4000003 ! single value
is_a: QC:4000009 ! ID based
is_a: QC:4000022 ! chromatogram metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
