% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{rtOverMSQuarters}
\alias{rtOverMSQuarters}
\title{MS1/MS2 quantiles RT fraction (QC:4000055/QC:4000056)}
\usage{
rtOverMSQuarters(spectra, MSLevel = 1L)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{MSLevel}{`numeric(1)`}
}
\value{
`numeric(4)`
}
\description{
"The interval used for acquisition of the first, second, third, and fourth 
quarter of all MS1 events divided by RT-Duration." [PSI:QC]
}
\details{
is_a: QC:4000004 ! n-tuple
is_a: QC:4000010 ! ID free
is_a: QC:4000021 ! retention time metric
is_a: QC:4000023 ! MS1 metric
}
\examples{

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
