% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_visualizations.R
\name{shinyMsQuality}
\alias{shinyMsQuality}
\title{Shiny application to visualize quality metrics}
\usage{
shinyMsQuality(qc)
}
\arguments{
\item{qc}{`data.frame`}
}
\value{
`shiny`
}
\description{
The function `shinyMsQuality` function starts a shiny application to 
visualize the quality metrics interactively. It allows to display all metrics
contained in `qc`. 

The function accepts the output of `calculateMetrics`,
`calculateMetricsFromSpectra`, or `calculateMetricsFromMsExperiment`
}
\details{
The plots within the shiny application can be saved by clicking on the 
download button.
}
\examples{
library(msdata)
library(MsExperiment)
library(S4Vectors)
mse <- MsExperiment()
sd <- DataFrame(sample_id = c("QC1", "QC2"),
    sample_name = c("QC Pool", "QC Pool"), injection_idx = c(1, 3))
sampleData(mse) <- sd

## define file names containing spectra data for the samples and
## add them, along with other arbitrary files to the experiment
fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
experimentFiles(mse) <- MsExperimentFiles(
    mzML_files = fls,
    annotations = "internal_standards.txt")
## link samples to data files: first sample to first file in "mzML_files",
## second sample to second file in "mzML_files"
mse <- linkSampleData(mse, with = "experimentFiles.mzML_files",
    sampleIndex = c(1, 2), withIndex = c(1, 2))
mse <- linkSampleData(mse, with = "experimentFiles.annotations",
                      sampleIndex = c(1, 2), withIndex = c(1, 1))

library(Spectra)
## import the data and add it to the mse object
spectra(mse) <- Spectra(fls, backend = MsBackendMzR())

## define the quality metrics to be calculated
metrics <- c("areaUnderTIC", "rtDuration", "msSignal10XChange")

## additional parameters passed to the quality metrics functions
## (MSLevel is an argument of areaUnderTIC and msSignal10XChange,
## relativeTo is an argument of msSignal10XChange)
params_l <- list(MSLevel = 1, relativeTo = c("Q1", "previous"), 
    change = c("jump", "fall"))
    
## calculate the metrics
qc <- calculateMetricsFromMsExperiment(msexp = mse, metrics = metrics, 
    params = params_l)
rownames(qc) <- c("Sample 1", "Sample 2")

\dontrun{
shinyMsQuality(qc)
}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
