% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Chromatograms_metrics.R
\name{signalToNoiseRatio}
\alias{signalToNoiseRatio}
\title{Signal-to-Noise Ratio across chromatograms}
\usage{
signalToNoiseRatio(chromatograms, ...)
}
\arguments{
\item{chromatograms}{`Chromatograms` object}

\item{...}{further arguments (currently ignored)}
}
\value{
`numeric(1)`
}
\description{
The function `signalToNoiseRatio` calculates a robust signal-to-noise ratio
approximated as the Maximum Intensity divided by the Median of Non-Zero
Intensities.
}
\details{
The noise is estimated as the median of all intensity values greater than 0.
This avoids underestimating noise in sparse chromatograms where the
5th percentile might be 0.
}
\author{
Philippine Louail
}
