% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Spectra_metrics.R
\name{ticQuantileToQuantileLogRatio}
\alias{ticQuantileToQuantileLogRatio}
\title{MS1 quantile TIC change ratio to Quantile 1 (QC:4000057) or to 
previous quantile (QC:4000058)}
\usage{
ticQuantileToQuantileLogRatio(spectra, relativeTo = "Q1", msLevel = 1L, ...)
}
\arguments{
\item{spectra}{`Spectra` object}

\item{relativeTo}{`character(1)`, one of `"Q1"` or `"previous"`}

\item{msLevel}{`integer`}

\item{...}{not used here}
}
\value{
`numeric(1)`
}
\description{
"The log ratio for the second to n-th quantile of TIC changes over first 
quantile of TIC changes." [PSI:QC]
id: QC:4000057

"The log ratio for the second to n-th quantile of TIC over the previous 
quantile of TIC. For the boundary elements min/max are used." [PSI:QC]
id: QC:4000058
}
\details{
is_a: QC:4000004 ! n-tuple
is_a: QC:4000010 ! ID free
is_a: QC:4000022 ! chromatogram metric
is_a: QC:4000023 ! MS1 metric

*TIC changes* are interpreted as follows:
(1) the cumulative sum (`cumsum`) of the  `spectra`'s TIC is calculated 
(with spectra ordered by retention time),

(2) quartiles are then calculated on these, 

(3) for *QC:4000057* the log2 ratio between the 25, 50, 75
and 100% quantile to the 0% quantile is calculated. For *QC:4000058* 
ratios between the 25/0, 50/25, 75/50 and 100/75% quantiles are calculated.

The `log2` values are returned instead of the `log` values.
}
\note{
This function interprets the *quantiles* from the [PSI:QC] definition as
*quartiles*, i.e. the 0, 25, 50, 75 and 100% quantiles are used.
}
\examples{
library(S4Vectors)
library(Spectra)

spd <- DataFrame(
    msLevel = c(2L, 2L, 2L),
    polarity = c(1L, 1L, 1L),
    id = c("HMDB0000001", "HMDB0000001", "HMDB0001847"),
    name = c("1-Methylhistidine", "1-Methylhistidine", "Caffeine"))
## Assign m/z and intensity values
spd$mz <- list(
    c(109.2, 124.2, 124.5, 170.16, 170.52),
    c(83.1, 96.12, 97.14, 109.14, 124.08, 125.1, 170.16),
    c(56.0494, 69.0447, 83.0603, 109.0395, 110.0712,
        111.0551, 123.0429, 138.0662, 195.0876))
spd$intensity <- list(
    c(3.407, 47.494, 3.094, 100.0, 13.240),
    c(6.685, 4.381, 3.022, 16.708, 100.0, 4.565, 40.643),
    c(0.459, 2.585, 2.446, 0.508, 8.968, 0.524, 0.974, 100.0, 40.994))
sps <- Spectra(spd)
ticQuantileToQuantileLogRatio(spectra = sps, relativeTo = "Q1",
    msLevel = 2L)
ticQuantileToQuantileLogRatio(spectra = sps, relativeTo = "previous",
    msLevel = 2L)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
